internal: add tags and stubs so wazero can be compiled for MCU targets without operating systems
