+++ internal/sysfs/datasync_tinygo.go
@@ -1,14 +1,13 @@
-//go:build linux && !tinygo
+//go:build tinygo
 
 package sysfs
 
 import (
 	"os"
-	"syscall"
 
 	"github.com/tetratelabs/wazero/experimental/sys"
 )
 
 func datasync(f *os.File) sys.Errno {
-	return sys.UnwrapOSError(syscall.Fdatasync(int(f.Fd())))
+	return sys.ENOSYS
 }
