+++ internal/sysfs/dirfs.go
@@ -63,18 +63,6 @@ func (d *dirFS) Mkdir(path string, perm fs.FileMode) (errno experimentalsys.Errn
 	return
 }
 
-// Chmod implements the same method as documented on sys.FS
-func (d *dirFS) Chmod(path string, perm fs.FileMode) experimentalsys.Errno {
-	err := os.Chmod(d.join(path), perm)
-	return experimentalsys.UnwrapOSError(err)
-}
-
-// Rename implements the same method as documented on sys.FS
-func (d *dirFS) Rename(from, to string) experimentalsys.Errno {
-	from, to = d.join(from), d.join(to)
-	return rename(from, to)
-}
-
 // Readlink implements the same method as documented on sys.FS
 func (d *dirFS) Readlink(path string) (string, experimentalsys.Errno) {
 	// Note: do not use syscall.Readlink as that causes race on Windows.
@@ -89,20 +77,6 @@ func (d *dirFS) Readlink(path string) (string, experimentalsys.Errno) {
 // Rmdir implements the same method as documented on sys.FS
 func (d *dirFS) Rmdir(path string) experimentalsys.Errno {
 	return rmdir(d.join(path))
-}
-
-// Unlink implements the same method as documented on sys.FS
-func (d *dirFS) Unlink(path string) (err experimentalsys.Errno) {
-	return unlink(d.join(path))
-}
-
-// Symlink implements the same method as documented on sys.FS
-func (d *dirFS) Symlink(oldName, link string) experimentalsys.Errno {
-	// Note: do not resolve `oldName` relative to this dirFS. The link result is always resolved
-	// when dereference the `link` on its usage (e.g. readlink, read, etc).
-	// https://github.com/bytecodealliance/cap-std/blob/v1.0.4/cap-std/src/fs/dir.rs#L404-L409
-	err := os.Symlink(oldName, d.join(link))
-	return experimentalsys.UnwrapOSError(err)
 }
 
 // Utimens implements the same method as documented on sys.FS
