+++ internal/sysfs/nonblock_unix.go
@@ -1,4 +1,4 @@
-//go:build !windows && !plan9
+//go:build !windows && !plan9 && !tinygo
 
 package sysfs
