+++ internal/sysfs/open_file_notwindows.go
@@ -1,4 +1,4 @@
-//go:build !windows
+//go:build !windows && !tinygo
 
 package sysfs
