+++ internal/sysfs/poll_linux.go
@@ -1,3 +1,5 @@
+//go:build !tinygo
+
 package sysfs
 
 import (
