+++ internal/sysfs/poll_unsupported.go
@@ -1,4 +1,4 @@
-//go:build !linux && !darwin && !windows
+//go:build !(linux || darwin || windows) || tinygo
 
 package sysfs
